+++ lib/Model/PrepaidReportsProductModel.php
@@ -1,0 +1,634 @@
+<?php
+
+namespace Secuconnect\Client\Model;
+
+/**
+ * PrepaidReportsProductModel
+ *
+ * @category Class
+ * @package  Secuconnect\Client
+ * @author   Swagger Codegen team
+ * @link     https://github.com/swagger-api/swagger-codegen
+ */
+class PrepaidReportsProductModel extends ProductInstanceUID
+{
+    const DISCRIMINATOR = null;
+
+    /**
+      * The original name of the model.
+      * @var string
+      */
+    protected static $swaggerModelName = 'PrepaidReportsProductModel';
+
+    /**
+      * Array of property to type mappings. Used for (de)serialization
+      * @var string[]
+      */
+    protected static $swaggerTypes = [
+        'l' => 'int',
+        'k' => 'int',
+        'ks' => 'string',
+        'c' => 'int',
+        's' => 'int',
+        't' => '\Secuconnect\Client\Model\AggregationTimeResult',
+        'created' => 'string',
+        'updated' => 'string',
+        'merchant' => '\Secuconnect\Client\Model\ProductInstanceUID',
+        'store' => '\Secuconnect\Client\Model\ProductInstanceUID',
+        'device' => '\Secuconnect\Client\Model\ProductInstanceUID',
+        'amount' => 'int',
+        'vtc_tid' => 'int',
+        'report_from' => 'string',
+        'report_to' => 'string',
+        'receipt' => '\Secuconnect\Client\Model\ReceiptTypeValue[]',
+        'receipt_header' => '\Secuconnect\Client\Model\ReceiptTypeValue[]'
+    ];
+
+    /**
+      * Array of property to format mappings. Used for (de)serialization
+      * @var string[]
+      */
+    protected static $swaggerFormats = [
+        'l' => null,
+        'k' => null,
+        'ks' => null,
+        'c' => null,
+        's' => null,
+        't' => null,
+        'created' => null,
+        'updated' => null,
+        'merchant' => null,
+        'store' => null,
+        'device' => null,
+        'amount' => null,
+        'vtc_tid' => null,
+        'report_from' => null,
+        'report_to' => null,
+        'receipt' => null,
+        'receipt_header' => null
+    ];
+
+    public static function swaggerTypes()
+    {
+        return self::$swaggerTypes + parent::swaggerTypes();
+    }
+
+    public static function swaggerFormats()
+    {
+        return self::$swaggerFormats + parent::swaggerFormats();
+    }
+
+    /**
+     * Array of attributes where the key is the local name, and the value is the original name
+     * @var string[]
+     */
+    protected static $attributeMap = [
+        'l' => 'l',
+        'k' => 'k',
+        'ks' => 'ks',
+        'c' => 'c',
+        's' => 's',
+        't' => 't',
+        'created' => 'created',
+        'updated' => 'updated',
+        'merchant' => 'merchant',
+        'store' => 'store',
+        'device' => 'device',
+        'amount' => 'amount',
+        'vtc_tid' => 'vtc_tid',
+        'report_from' => 'report_from',
+        'report_to' => 'report_to',
+        'receipt' => 'receipt',
+        'receipt_header' => 'receipt_header'
+    ];
+
+    /**
+     * Array of attributes to setter functions (for deserialization of responses)
+     * @var string[]
+     */
+    protected static $setters = [
+        'l' => 'setL',
+        'k' => 'setK',
+        'ks' => 'setKs',
+        'c' => 'setC',
+        's' => 'setS',
+        't' => 'setT',
+        'created' => 'setCreated',
+        'updated' => 'setUpdated',
+        'merchant' => 'setMerchant',
+        'store' => 'setStore',
+        'device' => 'setDevice',
+        'amount' => 'setAmount',
+        'vtc_tid' => 'setVtcTid',
+        'report_from' => 'setReportFrom',
+        'report_to' => 'setReportTo',
+        'receipt' => 'setReceipt',
+        'receipt_header' => 'setReceiptHeader'
+    ];
+
+    /**
+     * Array of attributes to getter functions (for serialization of requests)
+     * @var string[]
+     */
+    protected static $getters = [
+        'l' => 'getL',
+        'k' => 'getK',
+        'ks' => 'getKs',
+        'c' => 'getC',
+        's' => 'getS',
+        't' => 'getT',
+        'created' => 'getCreated',
+        'updated' => 'getUpdated',
+        'merchant' => 'getMerchant',
+        'store' => 'getStore',
+        'device' => 'getDevice',
+        'amount' => 'getAmount',
+        'vtc_tid' => 'getVtcTid',
+        'report_from' => 'getReportFrom',
+        'report_to' => 'getReportTo',
+        'receipt' => 'getReceipt',
+        'receipt_header' => 'getReceiptHeader'
+    ];
+
+    public static function attributeMap()
+    {
+        return parent::attributeMap() + self::$attributeMap;
+    }
+
+    public static function setters()
+    {
+        return parent::setters() + self::$setters;
+    }
+
+    public static function getters()
+    {
+        return parent::getters() + self::$getters;
+    }
+
+    /**
+     * Constructor
+     * @param mixed[] $data Associated array of property values initializing the model
+     */
+    public function __construct(array $data = null)
+    {
+        parent::__construct($data);
+
+        $this->container['l'] = isset($data['l']) ? $data['l'] : null;
+        $this->container['k'] = isset($data['k']) ? $data['k'] : null;
+        $this->container['ks'] = isset($data['ks']) ? $data['ks'] : null;
+        $this->container['c'] = isset($data['c']) ? $data['c'] : null;
+        $this->container['s'] = isset($data['s']) ? $data['s'] : null;
+        $this->container['t'] = isset($data['t']) ? $data['t'] : null;
+        $this->container['created'] = isset($data['created']) ? $data['created'] : null;
+        $this->container['updated'] = isset($data['updated']) ? $data['updated'] : null;
+        $this->container['merchant'] = isset($data['merchant']) ? $data['merchant'] : null;
+        $this->container['store'] = isset($data['store']) ? $data['store'] : null;
+        $this->container['device'] = isset($data['device']) ? $data['device'] : null;
+        $this->container['amount'] = isset($data['amount']) ? $data['amount'] : null;
+        $this->container['vtc_tid'] = isset($data['vtc_tid']) ? $data['vtc_tid'] : null;
+        $this->container['report_from'] = isset($data['report_from']) ? $data['report_from'] : null;
+        $this->container['report_to'] = isset($data['report_to']) ? $data['report_to'] : null;
+        $this->container['receipt'] = isset($data['receipt']) ? $data['receipt'] : null;
+        $this->container['receipt_header'] = isset($data['receipt_header']) ? $data['receipt_header'] : null;
+    }
+
+    /**
+     * show all the invalid properties with reasons.
+     *
+     * @return array invalid properties with reasons
+     */
+    public function listInvalidProperties()
+    {
+        $invalid_properties = parent::listInvalidProperties();
+
+        return $invalid_properties;
+    }
+
+    /**
+     * validate all the properties in the model
+     * return true if all passed
+     *
+     * @return bool True if all properties are valid
+     */
+    public function valid()
+    {
+        return true;
+    }
+
+
+    /**
+     * Gets l
+     * @return int
+     */
+    public function getL()
+    {
+        return $this->container['l'];
+    }
+
+    /**
+     * Sets l
+     * @param int $l lookup index
+     * @return $this
+     */
+    public function setL($l)
+    {
+        $this->container['l'] = $l;
+
+        return $this;
+    }
+
+    /**
+     * Gets k
+     * @return int
+     */
+    public function getK()
+    {
+        return $this->container['k'];
+    }
+
+    /**
+     * Sets k
+     * @param int $k key index
+     * @return $this
+     */
+    public function setK($k)
+    {
+        $this->container['k'] = $k;
+
+        return $this;
+    }
+
+    /**
+     * Gets ks
+     * @return string
+     */
+    public function getKs()
+    {
+        return $this->container['ks'];
+    }
+
+    /**
+     * Sets ks
+     * @param string $ks key name
+     * @return $this
+     */
+    public function setKs($ks)
+    {
+        $this->container['ks'] = $ks;
+
+        return $this;
+    }
+
+    /**
+     * Gets c
+     * @return int
+     */
+    public function getC()
+    {
+        return $this->container['c'];
+    }
+
+    /**
+     * Sets c
+     * @param int $c count (number of items)
+     * @return $this
+     */
+    public function setC($c)
+    {
+        $this->container['c'] = $c;
+
+        return $this;
+    }
+
+    /**
+     * Gets s
+     * @return int
+     */
+    public function getS()
+    {
+        return $this->container['s'];
+    }
+
+    /**
+     * Sets s
+     * @param int $s sum
+     * @return $this
+     */
+    public function setS($s)
+    {
+        $this->container['s'] = $s;
+
+        return $this;
+    }
+
+    /**
+     * Gets t
+     * @return \Secuconnect\Client\Model\AggregationTimeResult
+     */
+    public function getT()
+    {
+        return $this->container['t'];
+    }
+
+    /**
+     * Sets t
+     * @param \Secuconnect\Client\Model\AggregationTimeResult $t t
+     * @return $this
+     */
+    public function setT($t)
+    {
+        $this->container['t'] = $t;
+
+        return $this;
+    }
+
+    /**
+     * Gets created
+     * @return string
+     */
+    public function getCreated()
+    {
+        return $this->container['created'];
+    }
+
+    /**
+     * Sets created
+     * @param string $created created
+     * @return $this
+     */
+    public function setCreated($created)
+    {
+        $this->container['created'] = $created;
+
+        return $this;
+    }
+
+    /**
+     * Gets updated
+     * @return string
+     */
+    public function getUpdated()
+    {
+        return $this->container['updated'];
+    }
+
+    /**
+     * Sets updated
+     * @param string $updated updated
+     * @return $this
+     */
+    public function setUpdated($updated)
+    {
+        $this->container['updated'] = $updated;
+
+        return $this;
+    }
+
+    /**
+     * Gets merchant
+     * @return \Secuconnect\Client\Model\ProductInstanceUID
+     */
+    public function getMerchant()
+    {
+        return $this->container['merchant'];
+    }
+
+    /**
+     * Sets merchant
+     * @param \Secuconnect\Client\Model\ProductInstanceUID $merchant merchant
+     * @return $this
+     */
+    public function setMerchant($merchant)
+    {
+        $this->container['merchant'] = $merchant;
+
+        return $this;
+    }
+
+    /**
+     * Gets store
+     * @return \Secuconnect\Client\Model\ProductInstanceUID
+     */
+    public function getStore()
+    {
+        return $this->container['store'];
+    }
+
+    /**
+     * Sets store
+     * @param \Secuconnect\Client\Model\ProductInstanceUID $store store
+     * @return $this
+     */
+    public function setStore($store)
+    {
+        $this->container['store'] = $store;
+
+        return $this;
+    }
+
+    /**
+     * Gets device
+     * @return \Secuconnect\Client\Model\ProductInstanceUID
+     */
+    public function getDevice()
+    {
+        return $this->container['device'];
+    }
+
+    /**
+     * Sets device
+     * @param \Secuconnect\Client\Model\ProductInstanceUID $device device
+     * @return $this
+     */
+    public function setDevice($device)
+    {
+        $this->container['device'] = $device;
+
+        return $this;
+    }
+
+    /**
+     * Gets amount
+     * @return int
+     */
+    public function getAmount()
+    {
+        return $this->container['amount'];
+    }
+
+    /**
+     * Sets amount
+     * @param int $amount amount
+     * @return $this
+     */
+    public function setAmount($amount)
+    {
+        $this->container['amount'] = $amount;
+
+        return $this;
+    }
+
+    /**
+     * Gets vtc_tid
+     * @return int
+     */
+    public function getVtcTid()
+    {
+        return $this->container['vtc_tid'];
+    }
+
+    /**
+     * Sets vtc_tid
+     * @param int $vtc_tid vtc_tid
+     * @return $this
+     */
+    public function setVtcTid($vtc_tid)
+    {
+        $this->container['vtc_tid'] = $vtc_tid;
+
+        return $this;
+    }
+
+    /**
+     * Gets report_from
+     * @return string
+     */
+    public function getReportFrom()
+    {
+        return $this->container['report_from'];
+    }
+
+    /**
+     * Sets report_from
+     * @param string $report_from Report start date
+     * @return $this
+     */
+    public function setReportFrom($report_from)
+    {
+        $this->container['report_from'] = $report_from;
+
+        return $this;
+    }
+
+    /**
+     * Gets report_to
+     * @return string
+     */
+    public function getReportTo()
+    {
+        return $this->container['report_to'];
+    }
+
+    /**
+     * Sets report_to
+     * @param string $report_to Report end date
+     * @return $this
+     */
+    public function setReportTo($report_to)
+    {
+        $this->container['report_to'] = $report_to;
+
+        return $this;
+    }
+
+    /**
+     * Gets receipt
+     * @return \Secuconnect\Client\Model\ReceiptTypeValue[]
+     */
+    public function getReceipt()
+    {
+        return $this->container['receipt'];
+    }
+
+    /**
+     * Sets receipt
+     * @param \Secuconnect\Client\Model\ReceiptTypeValue[] $receipt Receipt header
+     * @return $this
+     */
+    public function setReceipt($receipt)
+    {
+        $this->container['receipt'] = $receipt;
+
+        return $this;
+    }
+
+    /**
+     * Gets receipt_header
+     * @return \Secuconnect\Client\Model\ReceiptTypeValue[]
+     */
+    public function getReceiptHeader()
+    {
+        return $this->container['receipt_header'];
+    }
+
+    /**
+     * Sets receipt_header
+     * @param \Secuconnect\Client\Model\ReceiptTypeValue[] $receipt_header Receipt
+     * @return $this
+     */
+    public function setReceiptHeader($receipt_header)
+    {
+        $this->container['receipt_header'] = $receipt_header;
+
+        return $this;
+    }
+
+    /**
+     * Returns true if offset exists. False otherwise.
+     * @param integer $offset Offset
+     * @return boolean
+     */
+    public function offsetExists($offset)
+    {
+        return isset($this->container[$offset]);
+    }
+
+    /**
+     * Gets offset.
+     * @param integer $offset Offset
+     * @return mixed
+     */
+    public function offsetGet($offset)
+    {
+        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+    }
+
+    /**
+     * Sets value based on offset.
+     * @param integer $offset Offset
+     * @param mixed   $value  Value to be set
+     * @return void
+     */
+    public function offsetSet($offset, $value)
+    {
+        if (is_null($offset)) {
+            $this->container[] = $value;
+        } else {
+            $this->container[$offset] = $value;
+        }
+    }
+
+    /**
+     * Unsets offset.
+     * @param integer $offset Offset
+     * @return void
+     */
+    public function offsetUnset($offset)
+    {
+        unset($this->container[$offset]);
+    }
+
+    /**
+     * Gets the string presentation of the object
+     * @return string
+     */
+    public function __toString()
+    {
+        if (defined('JSON_PRETTY_PRINT')) { // use JSON pretty print
+            return json_encode(\Secuconnect\Client\ObjectSerializer::sanitizeForSerialization($this), JSON_PRETTY_PRINT);
+        }
+
+        return json_encode(\Secuconnect\Client\ObjectSerializer::sanitizeForSerialization($this));
+    }
+}
+